Splits shared/websocket into ES modules
