+++ shared/websocket/src/logger.js
@@ -1,0 +1,5 @@
+export default debug => (...args) => {
+  if (debug) {
+    console.log(...args);
+  }
+};
